requirement.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-crispy-forms==1.8.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
